requirements.txt
flask
serpapi
langchain
langchain-groq
pandas
pydantic
python-dotenv